:recycle: Align utils.functions tests with the package split
functions.py became the functions/ package, so mirror it in the test tree:
rename test_functions.py to functions/test_json.py (the model_to_json tests),
and move the loop-helper tests out of test_utils.py into functions/test_loop.py.
No test behavior change.

Refs: https://github.com/ChanTsune/django-boost/issues/164
